+++ src/main/java/com/genius/gitget/global/security/service/JwtGenerator.java
@@ -10,8 +10,8 @@ import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
-@Component
 @Slf4j
+@Component
 public class JwtGenerator {
 
     public String generateAccessToken(final Key ACCESS_SECRET, final long ACCESS_EXPIRATION, User user) {
