+++ src/main/java/com/genius/gitget/global/security/service/JwtService.java
@@ -1,152 +1,25 @@
 package com.genius.gitget.global.security.service;
 
-import static com.genius.gitget.global.security.constants.JwtRule.ACCESS_PREFIX;
-import static com.genius.gitget.global.security.constants.JwtRule.JWT_ISSUE_HEADER;
-import static com.genius.gitget.global.security.constants.JwtRule.REFRESH_PREFIX;
-import static com.genius.gitget.global.util.exception.ErrorCode.JWT_TOKEN_NOT_FOUND;
-import static com.genius.gitget.global.util.exception.ErrorCode.NOT_AUTHENTICATED_USER;
-
-import com.genius.gitget.challenge.user.domain.Role;
 import com.genius.gitget.challenge.user.domain.User;
 import com.genius.gitget.global.security.constants.JwtRule;
-import com.genius.gitget.global.security.constants.TokenStatus;
-import com.genius.gitget.global.security.domain.Token;
-import com.genius.gitget.global.security.repository.TokenRepository;
-import com.genius.gitget.global.util.exception.BusinessException;
-import com.genius.gitget.global.util.exception.ErrorCode;
-import io.jsonwebtoken.Jwts;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.security.Key;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.ResponseCookie;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@Transactional(readOnly = true)
-@Slf4j
-public class JwtService {
-    private final CustomUserDetailsService customUserDetailsService;
-    private final JwtGenerator jwtGenerator;
-    private final JwtUtil jwtUtil;
-    private final TokenRepository tokenRepository;
+public interface JwtService {
+    String generateAccessToken(HttpServletResponse response, User user);
 
-    private final Key ACCESS_SECRET_KEY;
-    private final Key REFRESH_SECRET_KEY;
-    private final long ACCESS_EXPIRATION;
-    private final long REFRESH_EXPIRATION;
+    String generateRefreshToken(HttpServletResponse response, User user);
 
-    public JwtService(CustomUserDetailsService customUserDetailsService, JwtGenerator jwtGenerator,
-                      JwtUtil jwtUtil, TokenRepository tokenRepository,
-                      @Value("${jwt.access-secret}") String ACCESS_SECRET_KEY,
-                      @Value("${jwt.refresh-secret}") String REFRESH_SECRET_KEY,
-                      @Value("${jwt.access-expiration}") long ACCESS_EXPIRATION,
-                      @Value("${jwt.refresh-expiration}") long REFRESH_EXPIRATION) {
-        this.customUserDetailsService = customUserDetailsService;
-        this.jwtGenerator = jwtGenerator;
-        this.jwtUtil = jwtUtil;
-        this.tokenRepository = tokenRepository;
-        this.ACCESS_SECRET_KEY = jwtUtil.getSigningKey(ACCESS_SECRET_KEY);
-        this.REFRESH_SECRET_KEY = jwtUtil.getSigningKey(REFRESH_SECRET_KEY);
-        this.ACCESS_EXPIRATION = ACCESS_EXPIRATION;
-        this.REFRESH_EXPIRATION = REFRESH_EXPIRATION;
-    }
+    String resolveTokenFromCookie(HttpServletRequest request, JwtRule tokenPrefix);
 
-    public void validateUser(User requestUser) {
-        if (requestUser.getRole() == Role.NOT_REGISTERED) {
-            throw new BusinessException(NOT_AUTHENTICATED_USER);
-        }
-    }
+    String getIdentifierFromRefresh(String refreshToken);
 
-    public String generateAccessToken(HttpServletResponse response, User requestUser) {
-        String accessToken = jwtGenerator.generateAccessToken(ACCESS_SECRET_KEY, ACCESS_EXPIRATION, requestUser);
-        ResponseCookie cookie = setTokenToCookie(ACCESS_PREFIX.getValue(), accessToken, ACCESS_EXPIRATION / 1000);
-        response.addHeader(JWT_ISSUE_HEADER.getValue(), cookie.toString());
+    boolean validateAccessToken(String accessToken);
 
-        return accessToken;
-    }
+    boolean validateRefreshToken(String refreshToken, String identifier);
 
-    @Transactional
-    public String generateRefreshToken(HttpServletResponse response, User requestUser) {
-        String refreshToken = jwtGenerator.generateRefreshToken(REFRESH_SECRET_KEY, REFRESH_EXPIRATION, requestUser);
-        ResponseCookie cookie = setTokenToCookie(REFRESH_PREFIX.getValue(), refreshToken, REFRESH_EXPIRATION / 1000);
-        response.addHeader(JWT_ISSUE_HEADER.getValue(), cookie.toString());
+    void logout(HttpServletResponse response, String identifier);
 
-        tokenRepository.save(new Token(requestUser.getIdentifier(), refreshToken));
-        return refreshToken;
-    }
-
-    private ResponseCookie setTokenToCookie(String tokenPrefix, String token, long maxAgeSeconds) {
-        return ResponseCookie.from(tokenPrefix, token)
-                .path("/")
-                .maxAge(maxAgeSeconds)
-                .httpOnly(true)
-                .sameSite("None")
-                .secure(true)
-                .build();
-    }
-
-    public boolean validateAccessToken(String token) {
-        return jwtUtil.getTokenStatus(token, ACCESS_SECRET_KEY) == TokenStatus.AUTHENTICATED;
-    }
-
-    public boolean validateRefreshToken(String token, String identifier) {
-        boolean isRefreshValid = jwtUtil.getTokenStatus(token, REFRESH_SECRET_KEY) == TokenStatus.AUTHENTICATED;
-
-        Token storedToken = tokenRepository.findByIdentifier(identifier);
-        boolean isTokenMatched = storedToken.getToken().equals(token);
-
-        return isRefreshValid && isTokenMatched;
-    }
-
-    public String resolveTokenFromCookie(HttpServletRequest request, JwtRule tokenPrefix) {
-        Cookie[] cookies = request.getCookies();
-        if (cookies == null) {
-            throw new BusinessException(JWT_TOKEN_NOT_FOUND);
-        }
-        return jwtUtil.resolveTokenFromCookie(cookies, tokenPrefix);
-    }
-
-    public Authentication getAuthentication(String token) {
-        UserDetails principal = customUserDetailsService.loadUserByUsername(getUserPk(token, ACCESS_SECRET_KEY));
-        return new UsernamePasswordAuthenticationToken(principal, "", principal.getAuthorities());
-    }
-
-    private String getUserPk(String token, Key secretKey) {
-        return Jwts.parserBuilder()
-                .setSigningKey(secretKey)
-                .build()
-                .parseClaimsJws(token)
-                .getBody()
-                .getSubject();
-    }
-
-    public String getIdentifierFromRefresh(String refreshToken) {
-        try {
-            return Jwts.parserBuilder()
-                    .setSigningKey(REFRESH_SECRET_KEY)
-                    .build()
-                    .parseClaimsJws(refreshToken)
-                    .getBody()
-                    .getSubject();
-        } catch (Exception e) {
-            throw new BusinessException(ErrorCode.INVALID_JWT);
-        }
-    }
-
-    public void logout(User requestUser, HttpServletResponse response) {
-        tokenRepository.deleteById(requestUser.getIdentifier());
-
-        Cookie accessCookie = jwtUtil.resetToken(ACCESS_PREFIX);
-        Cookie refreshCookie = jwtUtil.resetToken(REFRESH_PREFIX);
-
-        response.addCookie(accessCookie);
-        response.addCookie(refreshCookie);
-    }
+    Authentication getAuthentication(String accessToken);
 }
