+++ src/main/java/com/genius/gitget/global/security/service/JwtServiceImpl.java
@@ -1,0 +1,154 @@
+package com.genius.gitget.global.security.service;
+
+import static com.genius.gitget.global.security.constants.JwtRule.ACCESS_PREFIX;
+import static com.genius.gitget.global.security.constants.JwtRule.JWT_ISSUE_HEADER;
+import static com.genius.gitget.global.security.constants.JwtRule.REFRESH_PREFIX;
+import static com.genius.gitget.global.util.exception.ErrorCode.JWT_TOKEN_NOT_FOUND;
+
+import com.genius.gitget.challenge.user.domain.User;
+import com.genius.gitget.global.security.constants.JwtRule;
+import com.genius.gitget.global.security.constants.TokenStatus;
+import com.genius.gitget.global.security.domain.Token;
+import com.genius.gitget.global.security.repository.TokenRepository;
+import com.genius.gitget.global.util.exception.BusinessException;
+import com.genius.gitget.global.util.exception.ErrorCode;
+import io.jsonwebtoken.Jwts;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.security.Key;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.ResponseCookie;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@Slf4j
+public class JwtServiceImpl implements JwtService {
+    private final CustomUserDetailsService customUserDetailsService;
+    private final TokenRepository tokenRepository;
+    private final JwtGenerator jwtGenerator;
+    private final JwtUtil jwtUtil;
+
+    private final Key ACCESS_SECRET_KEY;
+    private final Key REFRESH_SECRET_KEY;
+    private final long ACCESS_EXPIRATION;
+    private final long REFRESH_EXPIRATION;
+
+    public JwtServiceImpl(CustomUserDetailsService customUserDetailsService,
+                          TokenRepository tokenRepository,
+                          JwtGenerator jwtGenerator, JwtUtil jwtUtil,
+                          @Value("${jwt.access-secret}") String ACCESS_SECRET_KEY,
+                          @Value("${jwt.refresh-secret}") String REFRESH_SECRET_KEY,
+                          @Value("${jwt.access-expiration}") long ACCESS_EXPIRATION,
+                          @Value("${jwt.refresh-expiration}") long REFRESH_EXPIRATION) {
+        this.customUserDetailsService = customUserDetailsService;
+        this.tokenRepository = tokenRepository;
+        this.jwtGenerator = jwtGenerator;
+        this.jwtUtil = jwtUtil;
+        this.ACCESS_SECRET_KEY = jwtUtil.getSigningKey(ACCESS_SECRET_KEY);
+        this.REFRESH_SECRET_KEY = jwtUtil.getSigningKey(REFRESH_SECRET_KEY);
+        this.ACCESS_EXPIRATION = ACCESS_EXPIRATION;
+        this.REFRESH_EXPIRATION = REFRESH_EXPIRATION;
+    }
+
+
+    @Override
+    public String generateAccessToken(HttpServletResponse response, User requestUser) {
+        String accessToken = jwtGenerator.generateAccessToken(ACCESS_SECRET_KEY, ACCESS_EXPIRATION, requestUser);
+        ResponseCookie cookie = setTokenToCookie(ACCESS_PREFIX.getValue(), accessToken, ACCESS_EXPIRATION / 1000);
+        response.addHeader(JWT_ISSUE_HEADER.getValue(), cookie.toString());
+
+        return accessToken;
+    }
+
+    @Override
+    @Transactional
+    public String generateRefreshToken(HttpServletResponse response, User requestUser) {
+        String refreshToken = jwtGenerator.generateRefreshToken(REFRESH_SECRET_KEY, REFRESH_EXPIRATION, requestUser);
+        ResponseCookie cookie = setTokenToCookie(REFRESH_PREFIX.getValue(), refreshToken, REFRESH_EXPIRATION / 1000);
+        response.addHeader(JWT_ISSUE_HEADER.getValue(), cookie.toString());
+
+        tokenRepository.save(new Token(requestUser.getIdentifier(), refreshToken));
+        return refreshToken;
+    }
+
+    private ResponseCookie setTokenToCookie(String tokenPrefix, String token, long maxAgeSeconds) {
+        return ResponseCookie.from(tokenPrefix, token)
+                .path("/")
+                .maxAge(maxAgeSeconds)
+                .httpOnly(true)
+                .sameSite("Strict")
+                .secure(true)
+                .build();
+    }
+
+    @Override
+    public boolean validateAccessToken(String token) {
+        return jwtUtil.getTokenStatus(token, ACCESS_SECRET_KEY) == TokenStatus.AUTHENTICATED;
+    }
+
+    @Override
+    public boolean validateRefreshToken(String token, String identifier) {
+        boolean isRefreshValid = jwtUtil.getTokenStatus(token, REFRESH_SECRET_KEY) == TokenStatus.AUTHENTICATED;
+
+        Token storedToken = tokenRepository.findByIdentifier(identifier);
+        boolean isTokenMatched = storedToken.getToken().equals(token);
+
+        return isRefreshValid && isTokenMatched;
+    }
+
+    @Override
+    public String resolveTokenFromCookie(HttpServletRequest request, JwtRule tokenPrefix) {
+        Cookie[] cookies = request.getCookies();
+        if (cookies == null) {
+            throw new BusinessException(JWT_TOKEN_NOT_FOUND);
+        }
+        return jwtUtil.resolveTokenFromCookie(cookies, tokenPrefix);
+    }
+
+    @Override
+    public String getIdentifierFromRefresh(String refreshToken) {
+        try {
+            return Jwts.parserBuilder()
+                    .setSigningKey(REFRESH_SECRET_KEY)
+                    .build()
+                    .parseClaimsJws(refreshToken)
+                    .getBody()
+                    .getSubject();
+        } catch (Exception e) {
+            throw new BusinessException(ErrorCode.INVALID_JWT);
+        }
+    }
+
+    @Override
+    public Authentication getAuthentication(String token) {
+        UserDetails principal = customUserDetailsService.loadUserByUsername(getUserPk(token, ACCESS_SECRET_KEY));
+        return new UsernamePasswordAuthenticationToken(principal, "", principal.getAuthorities());
+    }
+
+    private String getUserPk(String token, Key secretKey) {
+        return Jwts.parserBuilder()
+                .setSigningKey(secretKey)
+                .build()
+                .parseClaimsJws(token)
+                .getBody()
+                .getSubject();
+    }
+
+    @Override
+    public void logout(HttpServletResponse response, String identifier) {
+        tokenRepository.deleteById(identifier);
+
+        Cookie accessCookie = jwtUtil.resetToken(ACCESS_PREFIX);
+        Cookie refreshCookie = jwtUtil.resetToken(REFRESH_PREFIX);
+
+        response.addCookie(accessCookie);
+        response.addCookie(refreshCookie);
+    }
+}
